Parameter Make._

How to handle "client", "shared" and "server" sections for top level structure items.

For shared and server, the boolean argument indicate if this declaration can lead to evaluation of a fragment.

val shared_str : bool -> Ppxlib.Parsetree.structure_item -> Ppxlib.Parsetree.structure_item list

val server_str : bool -> Ppxlib.Parsetree.structure_item -> Ppxlib.Parsetree.structure_item list

val client_str : Ppxlib.Parsetree.structure_item -> Ppxlib.Parsetree.structure_item list

How to handle "client", "shared" and "server" sections for top level signature items.

val shared_sig : Ppxlib.Parsetree.signature_item -> Ppxlib.Parsetree.signature_item list

val client_sig : Ppxlib.Parsetree.signature_item -> Ppxlib.Parsetree.signature_item list

val server_sig : Ppxlib.Parsetree.signature_item -> Ppxlib.Parsetree.signature_item list

val fragment : loc:Ppxlib.Location.t -> ?typ:Ppxlib.Parsetree.core_type -> context:Context.server -> num:string -> id:string Ppxlib.Location.loc -> unsafe:bool -> Ppxlib.Parsetree.expression -> Ppxlib.Parsetree.expression

How to handle "%client ..." and "%shared ..." expr.

val escape_inject : loc:Ppxlib.Location.t -> ?ident:string -> context:Context.escape_inject -> id:string Ppxlib.Location.loc -> unsafe:bool -> Ppxlib.Parsetree.expression -> Ppxlib.Parsetree.expression

How to handle escaped "~%ident" inside a fragment.

val prelude : Ppxlib.Location.t -> Ppxlib.Parsetree.structure

val postlude : Ppxlib.Location.t -> Ppxlib.Parsetree.structure